fix(models): change Withdrawn tyoe to float64

+++ internal/models/models.go
@@ -82,7 +82,7 @@ type HandleUserWithdrawRequest = struct {
 
 type Balance struct {
 	Current   float64 `json:"current"`
-	Withdrawn int     `json:"withdrawn"`
+	Withdrawn float64 `json:"withdrawn"`
 }
 
 type AccrualServiceResponse struct {
